use new GrammarTestCase to simplify tests
