+++ composer.json
@@ -84,7 +84,8 @@
     "autoload-dev": {
         "classmap": [
             "tests/TestCase.php",
-            "tests/DbTestCase.php"
+            "tests/DbTestCase.php",
+            "tests/GrammarTestCase.php"
         ],
         "psr-4": {
             "Winter\\Storm\\Tests\\": "tests/"
