+++ tests/Database/Schema/Grammars/PostgresSchemaGrammarTest.php
@@ -2,50 +2,23 @@
 
 namespace Winter\Storm\Tests\Database\Schema\Grammars;
 
-use Illuminate\Database\Connection;
-use Illuminate\Database\Query\Expression;
 use Illuminate\Database\Schema\Blueprint;
-use Illuminate\Database\Schema\Builder;
-use Illuminate\Database\Schema\ForeignIdColumnDefinition;
 use Winter\Storm\Database\Schema\Grammars\PostgresGrammar;
-use Mockery as m;
-use PHPUnit\Framework\TestCase;
 
-class PostgresSchemaGrammarTest extends TestCase
+class PostgresSchemaGrammarTest extends \GrammarTestCase
 {
-    protected function tearDown(): void
+    public function setUp(): void
     {
-        m::close();
-    }
+        parent::setUp();
 
-    protected function setupConnection(Blueprint $blueprint)
-    {
-        $connection = m::mock(Connection::class);
-        $connection->shouldReceive('getSchemaBuilder')->andReturn($this->getSchemaBuilder($blueprint));
-        return $connection;
-    }
-
-    protected function getSchemaBuilder(Blueprint $blueprint)
-    {
-        $schemaBuilder = m::mock(Builder::class);
-        $schemaBuilder->shouldReceive('getColumns')->andReturn($blueprint->getColumns());
-
-        return $schemaBuilder;
-    }
-
-    protected function runBlueprint(Blueprint $initialBlueprint, Blueprint $blueprint = null)
-    {
-        $connection = $this->setupConnection($initialBlueprint);
-        if (is_null($blueprint)) {
-            $blueprint = $initialBlueprint;
-        }
-        return $blueprint->toSql($connection, new PostgresGrammar);
+        $this->grammar = new PostgresGrammar;
     }
 
     public function testNoInitialModifiersAddNullable()
     {
         $initialBlueprint = new Blueprint('users');
         $initialBlueprint->string('name');
+        $this->setupConnection($initialBlueprint);
 
         $statements = $this->runBlueprint($initialBlueprint);
         $this->assertSame('alter table "users" add column "name" varchar(255) not null', $statements[0]);
@@ -53,7 +26,7 @@ class PostgresSchemaGrammarTest extends TestCase
         $changedBlueprint = new Blueprint('users');
         $changedBlueprint->string('name')->nullable()->change();
 
-        $statements = $this->runBlueprint($initialBlueprint, $changedBlueprint);
+        $statements = $this->runBlueprint($changedBlueprint);
         $parts = explode(', ', $statements[0]);
         $this->assertSame('alter table "users" alter column "name" type varchar(255)', $parts[0]);
         $this->assertSame('alter column "name" drop not null', $parts[1]);
@@ -63,6 +36,7 @@ class PostgresSchemaGrammarTest extends TestCase
     {
         $initialBlueprint = new Blueprint('users');
         $initialBlueprint->string('name')->nullable();
+        $this->setupConnection($initialBlueprint);
 
         $statements = $this->runBlueprint($initialBlueprint);
         $this->assertSame('alter table "users" add column "name" varchar(255) null', $statements[0]);
@@ -70,7 +44,7 @@ class PostgresSchemaGrammarTest extends TestCase
         $changedBlueprint = new Blueprint('users');
         $changedBlueprint->string('name')->default('admin')->change();
 
-        $statements = $this->runBlueprint($initialBlueprint, $changedBlueprint);
+        $statements = $this->runBlueprint($changedBlueprint);
         $parts = explode(', ', $statements[0]);
         $this->assertSame('alter table "users" alter column "name" type varchar(255)', $parts[0]);
         $this->assertSame('alter column "name"  null', $parts[1]);
@@ -81,6 +55,7 @@ class PostgresSchemaGrammarTest extends TestCase
     {
         $initialBlueprint = new Blueprint('users');
         $initialBlueprint->string('name')->nullable();
+        $this->setupConnection($initialBlueprint);
 
         $statements = $this->runBlueprint($initialBlueprint);
         $this->assertSame('alter table "users" add column "name" varchar(255) null', $statements[0]);
@@ -88,7 +63,7 @@ class PostgresSchemaGrammarTest extends TestCase
         $changedBlueprint = new Blueprint('users');
         $changedBlueprint->string('name')->default('admin')->nullable(false)->change();
 
-        $statements = $this->runBlueprint($initialBlueprint, $changedBlueprint);
+        $statements = $this->runBlueprint($changedBlueprint);
         $parts = explode(', ', $statements[0]);
         $this->assertSame('alter table "users" alter column "name" type varchar(255)', $parts[0]);
         $this->assertSame('alter column "name" set not null', $parts[1]);
