+++ tests/Database/Schema/Grammars/SqlServerSchemaGrammarTest.php
@@ -2,49 +2,23 @@
 
 namespace Winter\Storm\Tests\Database\Schema\Grammars;
 
-use Illuminate\Database\Connection;
-use Illuminate\Database\Query\Expression;
 use Illuminate\Database\Schema\Blueprint;
-use Illuminate\Database\Schema\ForeignIdColumnDefinition;
 use Winter\Storm\Database\Schema\Grammars\SqlServerGrammar;
-use Mockery as m;
-use PHPUnit\Framework\TestCase;
 
-class SqlServerSchemaGrammarTest extends TestCase
+class SqlServerSchemaGrammarTest extends \GrammarTestCase
 {
-    protected function tearDown(): void
+    public function setUp(): void
     {
-        m::close();
-    }
+        parent::setUp();
 
-    protected function setupConnection(Blueprint $blueprint)
-    {
-        $connection = m::mock(Connection::class);
-        $connection->shouldReceive('getSchemaBuilder')->andReturn($this->getSchemaBuilder($blueprint));
-        return $connection;
-    }
-
-    protected function getSchemaBuilder(Blueprint $blueprint)
-    {
-        $schemaBuilder = m::mock(Builder::class);
-        $schemaBuilder->shouldReceive('getColumns')->andReturn($blueprint->getColumns());
-
-        return $schemaBuilder;
-    }
-
-    protected function runBlueprint(Blueprint $initialBlueprint, Blueprint $blueprint = null)
-    {
-        $connection = $this->setupConnection($initialBlueprint);
-        if (is_null($blueprint)) {
-            $blueprint = $initialBlueprint;
-        }
-        return $blueprint->toSql($connection, new SqlServerGrammar);
+        $this->grammar = new SqlServerGrammar;
     }
 
     public function testNoInitialModifiersAddNullable()
     {
         $initialBlueprint = new Blueprint('users');
         $initialBlueprint->string('name');
+        $this->setupConnection($initialBlueprint);
 
         $statements = $this->runBlueprint($initialBlueprint);
         $this->assertSame('alter table "users" add "name" nvarchar(255) not null', $statements[0]);
@@ -52,7 +26,7 @@ class SqlServerSchemaGrammarTest extends TestCase
         $changedBlueprint = new Blueprint('users');
         $changedBlueprint->string('name')->nullable()->change();
 
-        $statements = $this->runBlueprint($initialBlueprint, $changedBlueprint);
+        $statements = $this->runBlueprint($changedBlueprint);
         $this->assertSame('alter table "users" alter column "name" nvarchar(255) null', $statements[1]);
     }
 
@@ -60,6 +34,7 @@ class SqlServerSchemaGrammarTest extends TestCase
     {
         $initialBlueprint = new Blueprint('users');
         $initialBlueprint->string('name')->nullable();
+        $this->setupConnection($initialBlueprint);
 
         $statements = $this->runBlueprint($initialBlueprint);
         $this->assertSame('alter table "users" add "name" nvarchar(255) null', $statements[0]);
@@ -67,7 +42,7 @@ class SqlServerSchemaGrammarTest extends TestCase
         $changedBlueprint = new Blueprint('users');
         $changedBlueprint->string('name')->default('admin')->change();
 
-        $statements = $this->runBlueprint($initialBlueprint, $changedBlueprint);
+        $statements = $this->runBlueprint($changedBlueprint);
         $this->assertSame('alter table "users" alter column "name" nvarchar(255) null', $statements[1]);
         $this->assertSame('alter table "users" add default \'admin\' for "name"', $statements[2]);
     }
@@ -76,6 +51,7 @@ class SqlServerSchemaGrammarTest extends TestCase
     {
         $initialBlueprint = new Blueprint('users');
         $initialBlueprint->string('name')->nullable();
+        $this->setupConnection($initialBlueprint);
 
         $statements = $this->runBlueprint($initialBlueprint);
         $this->assertSame('alter table "users" add "name" nvarchar(255) null', $statements[0]);
@@ -83,7 +59,7 @@ class SqlServerSchemaGrammarTest extends TestCase
         $changedBlueprint = new Blueprint('users');
         $changedBlueprint->string('name')->default('admin')->nullable(false)->change();
 
-        $statements = $this->runBlueprint($initialBlueprint, $changedBlueprint);
+        $statements = $this->runBlueprint($changedBlueprint);
         $this->assertSame('alter table "users" alter column "name" nvarchar(255) not null', $statements[1]);
         $this->assertSame('alter table "users" add default \'admin\' for "name"', $statements[2]);
     }
